requirements.txt
face-recognition==1.3.0
mysql-connector-python>=8.0.22
Pillow==8.0.1
requests==2.24.0
Jinja2